Prep work for UI refactor.
